Add spinner while displaying profile. Move code for profile from main file to separate module
