+++ src/index.js
@@ -2,15 +2,11 @@
 
 const program = require('commander');
 const opn = require('opn');
-const chalk = require('chalk');
 
 const crawler = require('./util/crawler');
 const countdown = require('./util/spinner');
 const prompt = require('./util/prompt');
-
-const log = console.log;
-const title = chalk.yellow;
-const body = chalk.green;
+const showProfile = require('./util/profile');
 
 let articles;
 
@@ -54,88 +50,20 @@ const showPostsByTimeline = (timeline) => {
 }
 
 /**
- * This is a function to display name and description of author.
- * @param {Object} profile - profile of the author
+ * This is a function to display the profile details of the author.
+ * @param {Object} profileInfo - profile of the author
  * @returns {null} null
  */
 
-const showNameAndDesc = (profile) => {
-    log(title('name'))
-    log(body(profile.name));
-    log(title('\ndescription'))
-    log(body(`${profile.desc}\n`));
-}
-
-/**
- * This is a function to display author profile details such as email,work,location,etc.
- * @param {Object} profile - profile of the author
- * @returns {null} null
- */
-
-const showUserDetails = (profile) => {
-    profile.field.forEach((field, index) => {
-        log(title(field));
-        log(body(`${profile.value[index]}\n`));
-    });
-}
-
-/**
- * This is a function to display the links of the author mentioned in profile.
- * @param {Object} profile - profile of the author
- * @returns {null} null
- */
-
-const showLinks = (profile) => {
-    log(title('links'));
-    profile.links.forEach(link => {
-        log(body(link));
-    });
-}
-
-/**
- * This is a function to display the details in sidebar such as skills/languages, projects and hacks, etc.
- * @param {Object} profile - profile of the author
- * @returns {null} null
- */
-
-const showSidebarDetails = (profile) => {
-    log('');
-    profile.sidebarHeader.forEach((header, index) => {
-        log(title(header));
-        log(body(`${profile.sidebarBody[index]}\n`));
-    });
-}
-
-/**
- * This is a function to display stats such as posts published, comments written, tags followed.
- * @param {Object} profile - profile of the author
- * @returns {null} null
- */
-
-const showStats = (profile) => {
-    log(title('stats'));
-    profile.stats.forEach(stat => {
-        log(body(stat));
-    });
-}
-
-/**
- * This is a driver function to call all the above functions that display profile details on terminal.
- * @param {Object} profile - profile of the author
- * @returns {null} null
- */
-
-const showUserProfile = (username) => {
-    crawler.fetchAuthorProfile(username).then((profile) => {
-        if (!profile.name) {
+const showAuthorProfile = (username) => {
+    countdown.start();
+    crawler.fetchAuthorProfile(username).then((profileInfo) => {
+        countdown.stop();
+        if (!profileInfo.name) {
             console.error("😱 User not found. Please try again.");
             process.exit(1);
         }
-        showNameAndDesc(profile);
-        showUserDetails(profile);
-        showLinks(profile);
-        showSidebarDetails(profile);
-        showStats(profile);
+        showProfile(profileInfo);
     });
 }
 
@@ -211,7 +139,7 @@ program
     .alias("a")
     .action((username, cmd) => {
         if (cmd.profile) {
-            showUserProfile(username);
+            showAuthorProfile(username);
         } else {
             countdown.start();
             crawler.fetchByAuthor(username).then(data => {
